Update PockonsoleV2.ino
Starting to build the framework for animations using the Easing Library.  New enums created to facilitate selection of a fade option.  I may change it so the actual variable is only used inside of the function because I don't want to worry about having to reset it.  Maybe I'll make a default option then also have the ability to have functions with private instances of the same type.